infra/slack_logic/requirements.txt
fastapi==0.115.0
langchain_community==0.3.1
langchain_core==0.3.6
langchain_huggingface==0.1.0
langchain_openai==0.2.1
langchain_qdrant==0.1.4
python-dotenv==1.0.1
qdrant_client==1.11.3
slack_bolt==1.20.1
slack_sdk==3.33.1
uvicorn==0.31.0
langchain_core
